backend/requirements.txt
# ── Core ──
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-multipart==0.0.12

# ── Supabase ──
supabase==2.10.0

# ── Background tasks ──
celery==5.4.0
redis==5.2.0

# ── Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ── CSV / Data ──
pandas==2.2.0

# ── HTTP client (for ReachInbox) ──
httpx==0.28.0

# ── AI API ──
google-generativeai==0.8.3
